pyproject.toml
[project]
name = "playwright-zchb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "2captcha-python>=1.5.1",
    "loguru>=0.7.3",
    "patchright>=1.52.5",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "playwright>=1.54.0",
    "httpx>=0.28.1",
]
